Update Readme With Mission Statement
In order to focus our development efforts, we should settle on a statement of the core nature and purpose of Science Island. I've drafted a statement that I will circulate to the rest of the group for discussion. I have also added a summary for the doc, but that will probably change as the doc evolves.